requirements.txt
astroid==1.6.2
Django==2.0.3
djangorestframework==3.7.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.3
pytz==2018.3
six==1.11.0
wrapt==1.10.11
